+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 45976,
-  "expoServerNgrokUrl": "https://8r-xp6.anonymous.shopu.exp.direct",
-  "packagerNgrokUrl": "https://packager.8r-xp6.anonymous.shopu.exp.direct",
-  "ngrokPid": 45993
+  "packagerPid": 49755,
+  "expoServerNgrokUrl": "https://h9-ysn.anonymous.shopu.exp.direct",
+  "packagerNgrokUrl": "https://packager.h9-ysn.anonymous.shopu.exp.direct",
+  "ngrokPid": 49774
 }
