+++ .expo/settings.json
@@ -3,5 +3,5 @@
   "lanType": "ip",
   "dev": true,
   "minify": false,
-  "urlRandomness": "8r-xp6"
+  "urlRandomness": "h9-ysn"
 }
